backend/requirements.txt
# Server
uvicorn[standard]~=0.38.0
fastapi~=0.119.1
pyjwt~=2.10.1

# Database
alembic~=1.17.0  # migrations
alembic-postgresql-enum~=1.8.0  # psql enum migrations
SQLAlchemy[asyncio]~=2.0.40  # ORM
asyncpg~=0.30.0  # driver
bcrypt~=5.0.0  # auth

# Payment
yookassa~=3.7.1

# Cache
redis[hiredis]~=7.0.0